fix minor UI and code issues
