+++ web/components/issues/issue-layouts/kanban/block.tsx
@@ -6,6 +6,7 @@ import { TIssue, IIssueDisplayProperties, IIssueMap } from "@plane/types";
 // hooks
 import { ControlLink, DropIndicator, TOAST_TYPE, Tooltip, setToast } from "@plane/ui";
 import RenderIfVisible from "@/components/core/render-if-visible-HOC";
+import { HIGHLIGHT_CLASS } from "@/components/issues/issue-layouts/utils";
 import { cn } from "@/helpers/common.helper";
 import { useApplication, useIssueDetail, useKanbanView, useProject } from "@/hooks/store";
 import useOutsideClickDetector from "@/hooks/use-outside-click-detector";
@@ -133,7 +134,7 @@ export const KanbanIssueBlock: React.FC<IssueBlockProps> = observer((props) => {
   const isDragAllowed = !isDragDisabled && !issue?.tempId && canEditIssueProperties;
 
   useOutsideClickDetector(cardRef, () => {
-    cardRef?.current?.classList?.remove("highlight");
+    cardRef?.current?.classList?.remove(HIGHLIGHT_CLASS);
   });
 
   // Make Issue block both as as Draggable and,
