+++ web/components/issues/issue-layouts/kanban/utils.ts
@@ -1,5 +1,4 @@
 import pull from "lodash/pull";
-import scrollIntoView from "smooth-scroll-into-view-if-needed";
 import { IPragmaticDropPayload, TIssue, TIssueGroupByOptions } from "@plane/types";
 import { ISSUE_FILTER_DEFAULT_DATA } from "@/store/issue/helpers/issue-helper.store";
 
@@ -211,19 +210,4 @@ export const handleDragDrop = async (
       }))
     );
   }
-};
-
-/**
- * This Method finds the DOM element with elementId, scrolls to it and highlights the issue block
- * @param elementId
- * @param shouldScrollIntoView
- */
-export const highlightIssueOnDrop = (elementId: string | undefined, shouldScrollIntoView = true) => {
-  setTimeout(async () => {
-    const sourceElementId = elementId ?? "";
-    const sourceElement = document.getElementById(sourceElementId);
-    sourceElement?.classList?.add("highlight");
-    if (shouldScrollIntoView && sourceElement)
-      await scrollIntoView(sourceElement, { behavior: "smooth", block: "center", duration: 1500 });
-  }, 200);
 };
